LayoutTests/fast/profiler/execution-context-and-eval-on-same-line.html

<html>
<head>
<script src="resources/profiler-test-JS-resources.js"></script>
<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();

console.profile("Two Execution Contexts on the same line");
function testStart()
{
    evalFunction();
    endTest();
}
</script>
<script> eval("function evalFunction() { insertNewText(); }");</script>
</head>

<body onload="testStart()">
This page has a call to eval and a script tag on the same line.
<br>
<br>
To run this test manually, load it in the browser then load the WebInspector and look at
the profile.  In the profile there should be multiple (program) elements and
one that represents the eval().
<div id="output"></div>
</body>
</html>
